move registering "accesses to self" into `JavaClasses`
We can then remove the whole `AccessContext` within `DomainObjectCreationContext`.
